Update Peer review
Flawless! 🎉

Some tiny winy changes you can make to the button:

First, add the line below to specify that the button is clickable:
.add-card {
  cursor: pointer;
}
This will change the pointer to a hand-like shape.

Second, you don't necessarily need to wrap the elements inside the <button class="add-cart"> within the <div class="button-wrapper". The <button> itself is considered a wrapper. Therefore, you can add the flex to the button itself.
Hats off to your solution!

😎😎😎